refactor: remove workerpool #43
