+++ server/package.json
@@ -30,7 +30,6 @@
     "@types/node": "^16.11.7",
     "@types/node-fetch": "^2.6.2",
     "@types/uuid": "^9.0.0",
-    "@types/workerpool": "^6.1.0",
     "@typescript-eslint/eslint-plugin": "5.45.0",
     "@typescript-eslint/parser": "5.45.0",
     "eslint": "8.28.0",
@@ -68,7 +67,6 @@
     "rxjs": "^7.5.7",
     "source-map-support": "^0.5.21",
     "swagger-ui-express": "^4.6.0",
-    "uuid": "^9.0.0",
-    "workerpool": "^6.3.1"
+    "uuid": "^9.0.0"
   }
 }
