+++ server/src/handler/index.ts
@@ -1,2 +1,3 @@
 export * from './server-image-handler'
 export * from './web-image-handler'
+export { default as processor } from './process-job'
