+++ server/src/service/index.ts
@@ -1,2 +1,1 @@
 export * from './job-service'
-export * from './worker-service'
