+++ server/src/util/index.ts
@@ -1,2 +1,4 @@
 export * from './is-number-string-range'
 export * from './get-pkg-dir'
+export * from './get-paths'
+export * from './upload-buffer-to-url'
