+++ server/src/util/upload-buffer-to-url.ts
@@ -1,0 +1,8 @@
+import fetch from 'node-fetch'
+
+export const uploadBufferToUrl = async (url: string, buffer: Buffer): Promise<void> => {
+  await fetch(url, {
+    method: 'PUT',
+    body: buffer,
+  })
+}
