--- server/src/worker/index.ts
@@ -1,1 +1,0 @@
-export * from './server-image-handler-worker'
