--- server/src/worker/server-image-handler-worker.ts
@@ -1,9 +1,0 @@
-import { worker } from 'workerpool'
-import { ServerImageHandler } from '../handler'
-
-const handler = new ServerImageHandler()
-
-worker({
-  ping: handler.ping,
-  handlePath: handler.handlePath,
-})
