+++ server/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "../tsconfig.base.json",
   "compilerOptions": {
     "rootDir": ".",
+    "baseUrl": "server",
     "outDir": "dist"
   },
   "include": ["src"]
